modules/storage/storage-driver-gcs/pyproject.toml
[tool.poetry]
name = "toku-storage-driver-gcs"
version = "0.1.0"
description = "Project that provides the Google Cloud Storage (GCS) implemention for operating with storage driver"
authors = ["Toku <dev@trytoku.com>"]
license = "Internal"
readme = "README.md"
homepage = "https://www.jira.com"
repository = "https://www.github.com"
documentation = "https://www.notion.com"
packages = [
    {include = "toku"},
    {include = "tests"}
]

[tool.poetry.dependencies]
python = "^3.11"
overrides = "==7.4.0"
google-cloud-storage = "==2.13.0"
google-auth = "==2.23.4"
toku-storage-driver-api = "==0.1.0"


[tool.poetry.group.dev.dependencies]
pylint = "==3.0.2"
flake8 = "==6.1.0"
mypy = "==1.7.0"
black = "==23.11.0"

[tool.poetry.group.test.dependencies]
pytest = "==7.4.3"
pytest-cov = "==4.1.0"
pytest-xdist = "==3.3.1"
flexmock = "==0.11.3"
faker = "==20.0.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
